Update setup.py
Watched some youtube clips on this and though I'd try this software
I got an error soon after installing - the error was because `shlex.join` is being used:
https://github.com/calebstewart/pwncat/blob/5f12a129681bbdb7cfaa822aa09610236f42eaba/pwncat/__main__.py#L29
Turns out this requires Python 3.8: 
https://docs.python.org/3/library/shlex.html#shlex.join
I propose to explicitly require this at install time so a more appropriate error is thrown to the user. 
Thoughts?